Added HomeFragment and HomeViewModel
Added HomeFragment, HomeViewModel. Also created UI and moved InjectorUtils to the utility package then removed the utilities package. Adjusted the DealerDao and ModelDao to return null since it is possible to not have any selected model or dealer. Created MockFragment to have as our main navigation and changed the MainActivity to load it.