Created MyEquipment screen (#24)
Updated fragment to display a list of my equipment. Also created the views for the cells, added images for the category of each item. Added a FAB to add equipment. Also fixed crash in PreferenceSyncService when parsing user preferences. Created new parser to parse the data we get from the server when asking for user preferences. Created ViewModel class for the MyEquipmentListFragment. Created class to help with swiping motion for the model preferences.