Created the Maintenance Guide Activity (#41)
* Created the Maintenance Guide Activity

Created the maintenance guide layouts. Added new theme and resources to style layout correctly. Added new drawables for layout. Made changes to repo to get the guide pages also added class to get the text for the guide page. Added Picasso SDK to load the images for the page and used MediaPlayer to play the audio file for the page.