Fixed crash on dealer locator (#85)
Crashed was caused due to clearing out the markers on the google map but still holding on to a cleared marker in the form of the lastClickedMarker variable.
Fixing the crash exposed unusual behavior with the bottom sheet dialog. Instead of dealing with the behavior state, hide and show the views appropriately.