Updated the MSAL library to 0.3.1-alpha (#111)
* Updated the MSAL library to 0.3.1-alpha

Adjusted how we look for MSAL account objects and also how to authenticate users as per new library specs.

* Fixed bug when migrating from version 1 to 2

Fixed migration bug where  it would not sync account when token is expired. Now if you are migrating from version 1 to 2, the user will be logged out and will need to re-log in but account will sync when token is expired.