Created SignIn Service (#31)
* Created SignIn Service

Created a service for sign in. Also created ViewModel for the SignInFragment as well as took care of the use case of refreshing the token. I also added the API URL, Secret and ClientId to the BuildConfig file via Gradle. I removed the MSAL library and code pertaining to it.